-sort the applications based on recently used ids
